less stupid mod layout, more doc examples, test doc in CI
