+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rsmt2"
-version = "0.13.1"
+version = "0.14.0"
 authors = ["Adrien Champion <adrien.champion@email.com>"]
 description = "Wrapper for SMT-LIB 2 compliant SMT solvers."
 documentation = "https://docs.rs/rsmt2"
